// types/logo.ts
export type RawLogo = {
  id: number;
  nama: string;
  kategori: string;
  gambar1: string;
  gambar2: string;
  gambar3: string;
  gambar4: string;
  gambar5: string;
  gambar6: string;
  gambar7: string;
  gambar8: string;
  gambar9: string;
  gambar10: string;
};

export type Logo = {
  id: number;
  nama: string;
  kategori: string;
  gambar: string[];
};